Moves some requirements to new class + slightly changes format
